Normalize joystick axis input to float range

+++ code/client/client.h
@@ -128,7 +128,7 @@ typedef struct {
 
 	int mouseDx[2], mouseDy[2];         // added to by mouse events
 	int mouseIndex;
-	int joystickAxis[MAX_JOYSTICK_AXIS];            // set by joystick events
+	float joystickAxis[MAX_JOYSTICK_AXIS];            // set by joystick events
 
 	// cgame communicates a few values to the client system
 	int cgameUserCmdValue;              // current weapon to add to usercmd_t
